Mod. ServerRequirementsChecker. Rename class
