+++ lib/Cleantalk/ApbctWP/Activator.php
@@ -3,7 +3,7 @@
 namespace Cleantalk\ApbctWP;
 
 use Cleantalk\ApbctWP\AdjustToEnvironmentModule\AdjustToEnvironmentHandler;
-use Cleantalk\ApbctWP\ServerChecker\ServerChecker;
+use Cleantalk\ApbctWP\ServerRequirementsChecker\ServerRequirementsChecker;
 use Cleantalk\ApbctWP\UpdatePlugin\DbTablesCreator;
 
 class Activator
